refactor(agent_backends): extract codex_auth submodule
Move Codex auth-material plumbing (CodexAuthMaterial / CodexAuthJson /
CodexAuthTokens types, load_codex_auth_material,
codex_account_id_from_access_token, codex_auth_path, and the
CODEX_DEFAULT_BASE_URL / CODEX_JWT_AUTH_CLAIM constants) into a
focused codex_auth submodule.

The codex_account_id_can_be_derived_from_chatgpt_access_token test
moves with its helper. The codex_cli_command source-introspection
test (which scans mod.rs for the .no_console_window() invariant)
loses load_codex_auth_material as its sentinel; the next function
in mod.rs now is filter_openai_models, so the sentinel is updated
to match. (This sentinel will move to discovery.rs along with
codex_cli_command in a follow-up commit.)

Cross-module visibility uses pub(super) so the gateway translation
layer keeps reading CodexAuthMaterial directly and chat-send
keeps calling load_codex_auth_material. No public-surface change.

cargo fmt, cargo clippy on the three CI crates with -Dwarnings, and
the agent_backends test suite (111 passed) all pass.
